pyproject.toml
[project]
name = "code-trajectory-mcp"
version = "0.1.0"
authors = [{ name = "SynTaek" }]
description = "A Code Trajectory MCP server for tracking code evolution and providing context to LLMs."
readme = "README.md"
license = { file = "LICENSE" }
classifiers = ["License :: OSI Approved :: MIT License"]
requires-python = ">=3.14"
dependencies = ["mcp", "gitpython", "watchdog", "pyrefly>=0.44.0"]

[project.urls]
Homepage = "https://github.com/SynTaek"

[project.scripts]
code-trajectory-mcp = "code_trajectory.server:main"

[dependency-groups]
dev = ["pytest>=9.0.1", "ruff>=0.14.7"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/code_trajectory"]
